requirements.txt
# Django and DRF
Django==4.2.19
djangorestframework==3.14.0
django-environ==0.11.2
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0

# Celery and Redis
celery==5.3.6
redis==5.0.1
django-celery-results==2.5.1
django-celery-beat==2.5.0

# API Documentation
drf-yasg==1.21.7
django-debug-toolbar==4.2.0

# Internationalization
django-modeltranslation==0.18.11

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Development
django-debug-toolbar==4.2.0
black==23.11.0
flake8==6.1.0
isort==5.12.0